fix(pages/Profile): removed pages/Profile and placed inside of pages
